hickey + lowy: one refs row, because a relation is one fact
Both lenses over the diff, run together as they are meant to be.

**Hickey, layer 2 (fragmentation) and layer 4 (duplicated derivation).**
`AfterRefs.tsx` was `SeeRefs.tsx` with one word changed — read the field,
resolve each id through `nodeNamed`, draw a labelled row — and the copy
FRAGMENTED something worse than it duplicated: the row's label and its testid
are facts about a RELATION, they already lived in `relation.ts`, and each
component spelled its own. Renaming `see` in the table would have left both
components saying the old word, with everything still compiling. So the relation
arrives as a value (`edges/EdgeRefs.tsx`, which `SeeRefs.tsx` moved into) and
the words come out of the one table, which gained the row's testid for the same
reason it already held the label. A third writable edge field is a row there and
nothing anywhere else.

**Lowy, §1 prior-encapsulation.** Two candidates for "a parallel receptacle",
both examined and neither taken, recorded because a reviewer will ask:

- *the in-place panel under a row* (`date/DatePicker.tsx` is the other one).
  What the two share is `PAST_CONTROLS` and "Escape stops here"; their contents
  have nothing in common. That is layout, not an axis — variable, not volatile —
  and a `RowPanel` shell would be a boundary around where a box sits.
- *walking a shortlist with the arrows* (`complete/completing.tsx`, `palette/`).
  The axes under it ARE already encapsulated and this panel plugs into all of
  them — `search/nodes.ts` for the reading, `search/cursor.ts` for the walk,
  `search/Result.tsx` for the row, `keys.ts` for which key. What is left is what
  each key MEANS, which `keys.ts`'s own docstring assigns to the surface by
  name: `take` runs a route in the palette and rewrites a line in a completion.
