+++ packages/web/src/client/NodeBody.tsx
@@ -42,7 +42,7 @@ import { createMemo, Show } from "solid-js"
 import { DocRef } from "./document/DocRef.tsx"
 import { plainLine } from "./note/preview.ts"
 import { Note } from "./Note.tsx"
-import { SeeRefs } from "./SeeRefs.tsx"
+import { EdgeRefs } from "./edges/EdgeRefs.tsx"
 import { TESTID } from "./testids.ts"
 import { ROW_NOTE } from "./touch.ts"
 
@@ -134,7 +134,7 @@ export function NodeBody(props: {
                   </div>
                 )}
               </Show>
-              <SeeRefs node={props.shows.node} onRemove={props.onUnsee} />
+              <EdgeRefs node={props.shows.node} relation="see" onRemove={props.onUnsee} />
             </div>
           </Show>
 
@@ -156,7 +156,7 @@ export function NodeBody(props: {
           />
         )}
       </Show>
-      <SeeRefs node={props.shows.node} onRemove={props.onUnsee} />
+      <EdgeRefs node={props.shows.node} relation="see" onRemove={props.onUnsee} />
       <Show when={docOf(props.shows)}>
         {(doc) => <DocRef file={doc()} inline />}
       </Show>
